Fixed the counting for critical operations as well as an issue with recreating the report files

+++ Models/TableModel.java
@@ -2,9 +2,9 @@ package Models;
 
 public class TableModel {
     public int size;
-    public int countAverage;
+    public double countAverage;
     public double countCoef;
-    public long timeAverage;
+    public double timeAverage;
     public double timeCoef;
 
     public TableModel() {
